agent/requirements.txt
psutil
requests
pydantic
pyyaml
